+++ include/graphics/SpriteBatch.h
@@ -120,7 +120,7 @@ namespace vorb {
             // Render Batches
             ui32 _bufUsage;
             ui32 _vao, _vbo;
-            i32 _glyphCapacity;
+            size_t _glyphCapacity;
             std::vector<SpriteBatchCall*> _batches;
             PtrRecycler<SpriteBatchCall> _batchRecycler;
